Reverts utils.py back to its original content; adds _generate_custom_response to http module; implements tests
